add support for /tmp/telemetry_initialized status file ( OFW-2227 OFW-2402 )
OFW-2227 : telemtry uplaod not happening
OFW-2402 : Telemetry process tainted afterfactory reset and telemetry enable.

[Description]
1. When both Device.X_LGI-COM_Telemetry.Enable and Device.X_LGI-COM_Telemetry.DCMConfigFileURL
   parameters are set in single dmcli or ACS request, reload was attempted instead of start.
2. When is DCMConfigFileURL set immediately after Enable, kill -12 signal to reload is sent
   before telemtery is completely up and was going to tainted state.

[Solution]
1. Modiifed code to avoid reload if telemetry has to start.
2. Added check for a file based flag before sending kill -12 signal to reload config file.

[Testing Performed]
Tested and veriifed by enabling telemetry from boot config file and dmcli.

+++ source/dmltad/cosa_telemetry_dml.c
@@ -184,7 +184,10 @@ Telemetry_SetParamStringValue(ANSC_HANDLE hInsContext, char* ParamName, char* pS
         AnscCopyString(pMyObject->DCMConfigFileURL, pString);
         if (pMyObject->Enable == TRUE)  // Reload the service only if the Telemetry feature is enabled
         {
-            tele_state = TELE_ST_RELOAD;
+            if (tele_state != TELE_ST_START)
+            {
+                tele_state = TELE_ST_RELOAD;
+            }
         }
         CcspTraceDebug(("%s Set value %s: Exit\n", __FUNCTION__, pString));
         return TRUE;
@@ -235,12 +238,19 @@ Telemetry_Commit
 #else
                 system("killall -9 telemetry2_0");
 #endif
+                if(0 != remove("/tmp/telemetry_initialized")){
+                    CcspTraceWarning(("%s Unable to remove telemetry initialization complete flag \n", __FUNCTION__));
+                }
                 break;
             }
 	    case TELE_ST_RELOAD :
             {
                 CcspTraceInfo(("%s: Reloading Telemetry service with new configuration file \n", __FUNCTION__));
-                system("killall -12 telemetry2_0");
+                if( access( "/tmp/telemetry_initialized", F_OK ) != -1)
+                {
+                    CcspTraceInfo(("%s: Send signal 12 to reload configuration file \n", __FUNCTION__));
+                    system("killall -12 telemetry2_0");
+                }
                 break;
             }
        }
